[#noissue] Move the transaction id parse test out of DefaultTraceContextTest
