+++ agent-module/profiler/src/test/java/com/navercorp/pinpoint/profiler/context/DefaultTraceContextTest.java
@@ -20,9 +20,7 @@ import com.navercorp.pinpoint.bootstrap.config.DefaultProfilerConfig;
 import com.navercorp.pinpoint.bootstrap.config.ProfilerConfig;
 import com.navercorp.pinpoint.bootstrap.context.Trace;
 import com.navercorp.pinpoint.bootstrap.context.TraceContext;
-import com.navercorp.pinpoint.bootstrap.context.TraceId;
 import com.navercorp.pinpoint.common.profiler.util.TransactionId;
-import com.navercorp.pinpoint.common.profiler.util.TransactionIdUtils;
 import com.navercorp.pinpoint.profiler.context.id.DefaultTraceId;
 import com.navercorp.pinpoint.profiler.context.id.DefaultTransactionCounter;
 import com.navercorp.pinpoint.profiler.context.id.IdGenerator;
@@ -31,8 +29,6 @@ import com.navercorp.pinpoint.profiler.context.module.DefaultApplicationContext;
 import com.navercorp.pinpoint.profiler.context.provider.sampler.SamplerConfig;
 import com.navercorp.pinpoint.profiler.sampler.CountingSamplerFactory;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.AutoClose;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
@@ -44,64 +40,44 @@ import static org.mockito.Mockito.spy;
  */
 public class DefaultTraceContextTest {
 
-    @AutoClose
-    private DefaultApplicationContext applicationContext;
-
-    @BeforeEach
-    public void setUp() throws Exception {
-        ProfilerConfig profilerConfig = getProfilerConfig();
-
-        applicationContext = MockTraceContextFactory.newMockApplicationContext(profilerConfig);
-        applicationContext.start();
-    }
-
-    @Test
-    public void parseTest() {
-        String agent = "test";
-        long agentStartTime = System.currentTimeMillis();
-        long agentTransactionCount = 10;
-        TraceId traceId = new DefaultTraceId(TransactionId.of(agent, agentStartTime, agentTransactionCount));
-
-        String id = traceId.getTransactionId();
-
-        TransactionId transactionid = TransactionIdUtils.parseTransactionId(id);
-
-        Assertions.assertEquals(agent, transactionid.getAgentId());
-        Assertions.assertEquals(agentStartTime, transactionid.getAgentStartTime());
-        Assertions.assertEquals(agentTransactionCount, transactionid.getTransactionSequence());
-    }
 
     @Test
     public void disableTrace() {
+        ProfilerConfig profilerConfig = getProfilerConfig();
 
-        TraceContext traceContext = applicationContext.getTraceContext();
-        Trace trace = traceContext.disableSampling();
-        Assertions.assertNotNull(trace);
-        Assertions.assertFalse(trace.canSampled());
+        try (DefaultApplicationContext applicationContext = MockTraceContextFactory.newMockApplicationContext(profilerConfig)) {
+            applicationContext.start();
 
-        traceContext.removeTraceObject();
+            TraceContext traceContext = applicationContext.getTraceContext();
+            Trace trace = traceContext.disableSampling();
+            Assertions.assertNotNull(trace);
+            Assertions.assertFalse(trace.canSampled());
+
+            traceContext.removeTraceObject();
+        }
     }
 
     @Test
     public void threadLocalBindTest() {
-
-        TraceContext traceContext = applicationContext.getTraceContext();
-        Assertions.assertNotNull(traceContext.newTraceObject());
-
         ProfilerConfig profilerConfig = getProfilerConfig();
+        try (DefaultApplicationContext applicationContext = MockTraceContextFactory.newMockApplicationContext(profilerConfig);
+             DefaultApplicationContext applicationContext2 = MockTraceContextFactory.newMockApplicationContext(profilerConfig)) {
+            applicationContext.start();
+            TraceContext traceContext = applicationContext.getTraceContext();
+            Assertions.assertNotNull(traceContext.newTraceObject());
 
-        DefaultApplicationContext applicationContext2 = MockTraceContextFactory.newMockApplicationContext(profilerConfig);
-        applicationContext2.start();
+            applicationContext2.start();
 
-        TraceContext traceContext2 = applicationContext2.getTraceContext();
-        Trace notExist = traceContext2.currentRawTraceObject();
-        applicationContext2.close();
+            TraceContext traceContext2 = applicationContext2.getTraceContext();
+            Trace notExist = traceContext2.currentRawTraceObject();
 
-        Assertions.assertNull(notExist);
 
-        Assertions.assertNotNull(traceContext.currentRawTraceObject());
-        traceContext.removeTraceObject();
-        Assertions.assertNull(traceContext.currentRawTraceObject());
+            Assertions.assertNull(notExist);
+
+            Assertions.assertNotNull(traceContext.currentRawTraceObject());
+            traceContext.removeTraceObject();
+            Assertions.assertNull(traceContext.currentRawTraceObject());
+        }
     }
 
     @Test
@@ -113,44 +89,43 @@ public class DefaultTraceContextTest {
         Mockito.when((profilerConfig.readInt(CountingSamplerFactory.LEGACY_SAMPLING_RATE_NAME, -1))).thenReturn(samplingRate);
         Mockito.when((profilerConfig.readBoolean(SamplerConfig.SAMPLER_ENABLE_NAME, true))).thenReturn(true);
 
-        DefaultApplicationContext customContext = MockTraceContextFactory.newMockApplicationContext(profilerConfig);
-        customContext.start();
+        try (DefaultApplicationContext customContext = MockTraceContextFactory.newMockApplicationContext(profilerConfig)) {
+            customContext.start();
 
-        final TraceContext traceContext = customContext.getTraceContext();
-        IdGenerator idGenerator = customContext.getInjector().getInstance(IdGenerator.class);
-        final TransactionCounter transactionCounter = new DefaultTransactionCounter(idGenerator);
+            final TraceContext traceContext = customContext.getTraceContext();
+            IdGenerator idGenerator = customContext.getInjector().getInstance(IdGenerator.class);
+            final TransactionCounter transactionCounter = new DefaultTransactionCounter(idGenerator);
 
 
-        final long newTransactionCount = 22L;
-        @SuppressWarnings("unused")
-        final long expectedSampledNewCount = newTransactionCount / samplingRate + (newTransactionCount % samplingRate > 0 ? 1 : 0);
-        final long expectedUnsampledNewCount = newTransactionCount - expectedSampledNewCount;
-        for (int i = 0; i < newTransactionCount; i++) {
-            traceContext.newTraceObject();
-            traceContext.removeTraceObject();
+            final long newTransactionCount = 12;
+            @SuppressWarnings("unused")
+            final long expectedSampledNewCount = newTransactionCount / samplingRate + (newTransactionCount % samplingRate > 0 ? 1 : 0);
+            final long expectedUnsampledNewCount = newTransactionCount - expectedSampledNewCount;
+            for (int i = 0; i < newTransactionCount; i++) {
+                traceContext.newTraceObject();
+                traceContext.removeTraceObject();
+            }
+
+            final long expectedSampledContinuationCount = 5L;
+            for (int i = 0; i < expectedSampledContinuationCount; i++) {
+                traceContext.continueTraceObject(new DefaultTraceId(TransactionId.of("agentId", 0L, i)));
+                traceContext.removeTraceObject();
+            }
+
+            final long expectedUnsampledContinuationCount = 10L;
+            for (int i = 0; i < expectedUnsampledContinuationCount; i++) {
+                traceContext.disableSampling();
+                traceContext.removeTraceObject();
+            }
+
+            final long expectedTotalTransactionCount = expectedSampledNewCount + expectedUnsampledNewCount + expectedSampledContinuationCount + expectedUnsampledContinuationCount;
+
+            Assertions.assertEquals(expectedSampledNewCount, transactionCounter.getSampledNewCount());
+            Assertions.assertEquals(expectedUnsampledNewCount, transactionCounter.getUnSampledNewCount());
+            Assertions.assertEquals(expectedSampledContinuationCount, transactionCounter.getSampledContinuationCount());
+            Assertions.assertEquals(expectedUnsampledContinuationCount, transactionCounter.getUnSampledContinuationCount());
+            Assertions.assertEquals(expectedTotalTransactionCount, transactionCounter.getTotalTransactionCount());
         }
-
-        final long expectedSampledContinuationCount = 5L;
-        for (int i = 0; i < expectedSampledContinuationCount; i++) {
-            traceContext.continueTraceObject(new DefaultTraceId(TransactionId.of("agentId", 0L, i)));
-            traceContext.removeTraceObject();
-        }
-
-        final long expectedUnsampledContinuationCount = 10L;
-        for (int i = 0; i < expectedUnsampledContinuationCount; i++) {
-            traceContext.disableSampling();
-            traceContext.removeTraceObject();
-        }
-
-        customContext.close();
-
-        final long expectedTotalTransactionCount = expectedSampledNewCount + expectedUnsampledNewCount + expectedSampledContinuationCount + expectedUnsampledContinuationCount;
-
-        Assertions.assertEquals(expectedSampledNewCount, transactionCounter.getSampledNewCount());
-        Assertions.assertEquals(expectedUnsampledNewCount, transactionCounter.getUnSampledNewCount());
-        Assertions.assertEquals(expectedSampledContinuationCount, transactionCounter.getSampledContinuationCount());
-        Assertions.assertEquals(expectedUnsampledContinuationCount, transactionCounter.getUnSampledContinuationCount());
-        Assertions.assertEquals(expectedTotalTransactionCount, transactionCounter.getTotalTransactionCount());
     }
 
     public ProfilerConfig getProfilerConfig() {
